pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-notifier"
version = "0.1.0"
description = "Feishu bot for daily paper summaries"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apscheduler==3.10.4",
    "feedparser==6.0.11",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
